Reorganize constants
